+++ lib/main.dart
@@ -9,7 +9,6 @@ import 'package:txdx/screens/home_screen.dart';
 import 'package:txdx/screens/settings_screen.dart';
 import 'package:window_size/window_size.dart';
 
-import 'config/filters.dart';
 import 'providers/items/item_count_provider.dart';
 import 'providers/settings/settings_provider.dart';
 import 'config/settings.dart';
@@ -76,8 +75,8 @@ class TxDxApp extends ConsumerWidget {
   Widget build(BuildContext context, WidgetRef ref) {
 
     if (Platform.isMacOS) {
-      ref.listen(itemsCount(filterToday), (_, int? next) {
-        (next == null || next == 0)
+      ref.listen(badgeCount, (_, int next) {
+        (next == 0)
             ? FlutterAppBadger.removeBadge()
             : FlutterAppBadger.updateBadgeCount(next);
       });